trainModel/requirements.txt
python-dotenv
roboflow
ultralytics